Transaction_monitoring/currency/serializers.py
from rest_framework import serializers, viewsets

from currency.models import Currency


class CurrencySerializer(serializers.ModelSerializer):
    class Meta:
        model = Currency
        fields = ['id', 'code', 'name', 'symbol', 'exchange_rate_to_usd']
