Transaction_monitoring/requirements.txt
aio-pika==9.4.2
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiokafka==0.11.0
aiormq==6.8.0
aiosignal==1.3.1
aiosmtplib==3.0.2
aniso8601==9.0.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickhouse-connect==0.7.17
clickhouse-driver==0.2.8
coloredlogs==15.0.1
confluent-kafka==2.5.0
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==4.2.14
django-redis==5.4.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
dnspython==2.6.1
Flask==3.0.3
Flask-RESTful==0.3.10
frozenlist==1.4.1
geoip2==4.8.0
humanfriendly==10.0
idna==3.7
importlib_metadata==8.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
lz4==4.3.3
MarkupSafe==2.1.5
maxminddb==2.6.2
multidict==6.0.5
oauthlib==3.2.2
packaging==24.1
pamqp==3.3.0
pika==1.3.2
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pymongo==4.8.0
python3-openid==3.2.0
pytz==2024.1
redis==5.0.7
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2
zstandard==0.23.0
pymongo==4.8.0
fastapi==0.112.1
uvicorn==0.30.6
python-multipart==0.0.9
drf-spectacular==0.27.2
python-dateutil==2.9.0
httpx==0.27.0